mpa_merge: import occupancy_grid_utils
Upstream for occupancy_grid_utils is basically dead. Guys from clearpathrobotics have their repo at git@github.com:clearpathrobotics/occupancy_grid_utils.git. But they want to get rid of this package, so merging PRs is not their top priority. As Feb 2016 occupancy_grid_utils are not included in ROS distribution.

Lot of fuctionaly in occupancy_grid_utils is now obsolete as it is included in ROS core navigation stack.

* raytracing - costmap2d does this
* shortest paths - navfn and other planners have this
* loading from file - also something in navigation

The only useful functionaly not covered by navigation stack is imho *combineGrids*. I'm planning to do some improvements to this functionality (mostly focused on needs of map_merge). When the API (and ABI) stabilizes this may be splitted to it's own package (to make a full roundtrip :). Also I want to switch to C++11.

changes from POV of map_merge:

* imported parts from occupancy_grid_utils connected to combineGrids
* no changes for map_merge, still uses occupancy_grid_utils as library. Transition changes will follow.