added clang-format specification for project
this specifies code style for project. It differs from ROS recomendation in 2 ways

* not always breaks before braces (as this project uses this style everywhere)
* uses only 80 charecters per line

Primary ROS recomendation is to be consistent. This is the right style for me as the style changes will be applied gradually (at the right time, when files are refactored)

Formatting changes will be applied only for new files (strictly) and during major refactoring.