Update frontier_search.cpp
If there is obstacles between robot and fronter, then |p1-p2| can't represent the true distance. We can introduce the search depth into the distance, because the near frontier will be found at early iteration.  So we simply add a distance_cnt * resolution to the original distance to make it close to the true distance between robot and frontier in map.